Fix domain FromSql::accepts implementations
Closes sfackler/postgres#482
