+++ circle.yml
@@ -1,9 +1,8 @@
 version: 2
 jobs:
   build:
-    working_directory: ~/build
     docker:
-      - image: rust:1.20.0
+      - image: rust:1.31.0
       - image: postgres:9.6
         environment:
           POSTGRES_PASSWORD: password
